+++ modules/admin/src/commands/views/aw/classfile.php
@@ -45,7 +45,7 @@ class <?= $className; ?> extends ActiveWindow
     public function index()
     {
         return $this->render('index', [
-            'id' => $this->itemId,
+            'model' => $this->model,
         ]);
     }
 }