+++ modules/admin/src/commands/views/aw/viewfile.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @var $className
+ * @var $moduleId
+ */
+
+echo "<?php\n";
+?>
+
+/**
+ * <?= $className; ?> Index View.
+ *
+ * @var $this \luya\admin\ngrest\base\ActiveWindowView
+ * @var $model \luya\admin\ngrest\base\NgRestModel
+ */
+?>
+<h1>Active Window</h1>