+++ modules/admin/src/ngrest/base/ActiveWindowView.php
@@ -5,6 +5,8 @@ namespace luya\admin\ngrest\base;
 /**
  * Active Window View.
  *
+ * @property \luya\admin\ngrest\base\ActiveWindow $context
+ * 
  * @author Basil Suter <basil@nadar.io>
  */
 class ActiveWindowView extends \yii\base\View